Implement macOS accessibility support for terminal viewport
- Add ghostty_surface_viewport_text C API function to extract terminal viewport text
- Implement comprehensive NSAccessibility protocol in SurfaceView_AppKit.swift
- Add throttled accessibility notifications to avoid overwhelming the system
- Extract visible text from terminal cells with proper UTF-8 encoding
- Support for line breaks to maintain terminal text structure

This makes the terminal content readable by VoiceOver and other assistive technologies on macOS.

+++ macos/Sources/Ghostty/SurfaceView_AppKit.swift
@@ -1786,7 +1786,7 @@ extension Ghostty.SurfaceView {
         let buffer = UnsafeMutablePointer<CChar>.allocate(capacity: bufferSize)
         defer { buffer.deallocate() }
         
-        let length = ghostty_surface_viewport_text(surface, buffer, bufferSize)
+        let length = ghostty_surface_viewport_text(surface, buffer, UInt(bufferSize))
         guard length > 0 else { return "" }
         
         return String(cString: buffer)
@@ -1809,7 +1809,7 @@ extension Ghostty.SurfaceView {
         let buffer = UnsafeMutablePointer<CChar>.allocate(capacity: bufferSize)
         defer { buffer.deallocate() }
         
-        let length = ghostty_surface_selection(surface, buffer, bufferSize)
+        let length = ghostty_surface_selection(surface, buffer, UInt(bufferSize))
         guard length > 0 else { return nil }
         
         return String(bytesNoCopy: buffer, length: Int(length), encoding: .utf8, freeWhenDone: false)
